Fixing the Unit test which were failing of asset loading.
