+++ src/test/java/org/terasology/namegenerator/RegionNameProviderTest.java
@@ -18,43 +18,20 @@ package org.terasology.namegenerator;
 
 import static org.junit.Assert.assertFalse;
 
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+
 import org.junit.Test;
-import org.terasology.HeadlessEnvironment;
 import org.terasology.namegenerator.region.RegionNameProvider;
-import org.terasology.naming.Name;
 
 /**
  * Tests {@link RegionNameProvider}
  */
-public class RegionNameProviderTest {
-
-    private static HeadlessEnvironment env;
-
-    /**
-     * Setup headless environment
-     */
-    @BeforeClass
-    public static void setUpClass() {
-        env = new HeadlessEnvironment(new Name("NameGenerator"));
-    }
-
-    /**
-     * Clean up
-     * @throws Exception never
-     */
-    @AfterClass
-    public static void tearDownClass() throws Exception {
-        env.close();
-    }
+public class RegionNameProviderTest extends NameGeneratorTestingEnvironment {
 
     /**
      * Requires that original training data names do <b>NOT</b> contain any spaces
      */
     @Test
     public void testBase() {
-
         RegionNameProvider prov = new RegionNameProvider(123455);
 
         for (int i = 0; i < 100; i++) {
